backend/requirements.txt
fastapi
uvicorn
sqlalchemy
asyncpg
pydantic-settings
passlib
bcrypt==3.2.2
python-jose[cryptography]
alembic
langgraph
langchain-anthropic
redis
pgvector
psycopg2-binary
pytest
pytest-cov
pytest-asyncio
email-validator
python-multipart
httpx
asgi-lifespan
pip-audit
